Moving test to integration test
Also: setting some progress bars to False.
